requirements.txt
einops==0.3.0
fsspec==2023.9.0
ftfy==6.2.0
matplotlib==3.7.2
myutils==0.0.21
nltk==3.8.1
opencv_python==4.6.0.66
opencv_python_headless==4.8.0.76
openpyxl==3.1.2
Pillow==10.4.0
pycocotools==2.0.7
regex==2023.8.8
safetensors==0.4.3
scipy==1.14.0
numpy==1.26.4
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.1
transformers==4.51.0
timm==1.0.15
huggingface_hub
scikit-image
scikit-learn
open_clip_torch
PrettyTable
openmim
mmengine